+++ nodesdk/src/example/example2.ts
@@ -2,9 +2,6 @@ import { createJabraApplication } from '../main/index';
 
 createJabraApplication('A7tSsfD42VenLagL2mM6i2f0VafP/842cbuPCnC+uE8=').then((jabra) => {
     jabra.on('attach', (device) => {
-        // If you are creating a softphone, consider using GN protocol when device supports it.
-        // E.g. device.setHidWorkingStateAsync(enumHidState.GN_HID);
-
         device.isRingerSupportedAsync().then( (supported) => {
             if (supported) {
               device.offhookAsync().then ( () => {
@@ -30,7 +27,7 @@ createJabraApplication('A7tSsfD42VenLagL2mM6i2f0VafP/842cbuPCnC+uE8=').then((jab
 
         // In this demo example, we will auto shutdown once all jabra devices are removed:
         let remainingAttachedDevices = jabra.getAttachedDevices();
-        if (remainingAttachedDevices.size == 0) {
+        if (remainingAttachedDevices.length == 0) {
             jabra.disposeAsync().then (() => {}); // Cleanup and allow node process to exit.
         }
     });